Fix up Makefile.PL, restructure and rename.
We rename Krb5_Admin::* to Krb5Admin::* because this seems to be a
little more natural in Perl.  We also try to get the Makefile.PLs
into shape so that the build is a little more automagic and a bit
less manual.  We split the C bits off from the Perl bits which seems
to also help the structure out a bit.