Test for correct key being set.
Now that we've changed the key from des-cbc-crc to aes128-cts and
changed the key itself that we set in one of our tests, it makes
some level of sense to test that we in fact set the new key.  If
we continue to test that we set the old key which we are not in
fact setting any more then this test will continue to fail until
we make the fix that this commit actually is.