Initial attempt at Heimdal support.
These patches are the start of Heimdal support for these libraries.
Nico Williams <nico@cryptonector.com> provided the patches and I
modified them to some degree.  This is a work in progress, the
resulting library still does not pass all of the unit tests and
to get as far as it does, it requires modifications to Heimdal
which are not yet in the latest release.