Added a ticket placement database.
We add a number of methods to manage a ticket placement database
which is currently implemented in sqlite3.  The basic DB structure
defines a number of hosts and logical host mappings as well as
a list of ticket placements which can be either to a physical
host or a logical host.  Logical hosts will be expanded before
placement.

We also add a basic unit test for the functionality.