Make krb5_{set,rand}pass() in Krb5Admin::C take ks_tuples.
This changes the API a bit as we now pass in an array of integers
each of which represents an enctype.  We do not allow setting of
the salt type---we default to normal.  Krb5Admin::KerberosDB is
updated to use the new function signatures and the test cases are
updated as well.