Provide for xrealm bootstrapping.
To provide for xrealm bootstrapping, that is getting one's host keys
from realm A by demonstrating that you have host keys in realm B, we
define a configuration file (by default /etc/krb5/krb5_admind.conf)
which is a parsed Perl snippet and can define %xrealm_bootstrap.
This hash has keys which are the realm B (above) and the value is
a list of realm As.  E.g.:

%xrealm_bootstrap = (
	REALM1	=> ['REALM2', 'REALM3'],
);

means that if a client authenticates as host/foo@REALM1 then krb5_admind
will vend host keys for host/foo@REALM2 and host/foo@REALM3.