Lowercase winders princs before comparisons and make @v into $host_ok.
Windows principals are case insensitive and so when we're looking
at %win_xrealm_bootstrap, we need to be able to ingore case on the
lookups.  Rather than ignore case, we explicitly lower the case of
the client principal's components (but not realm) before we perform
the lookup and by doing so, we define the format of the config file
to specify only lowercase in principal names.

While we're here, we make the code a tad more explicit by removing
@v and replacing it with $host_ok which implies what we're doing a
little more clearly.

Patch by: viktor@dukhovni.org
Reviewed by: elric@imrryr.org