Update the prestashed ticket protocol and client to be realm aware.
We update krb5_prestash to connect to the correct KDCs for a realm
based on what it is doing.  We also extend
Krb5Admin::KerberosDB::fetch_tickets() to take a realm argument
which limits its results to that realm so that krb5_prestash can
make such a limitation in its requests.