Must check for undef before calling krb5_parse_name() in certain cases.
Looks like swig doesn't mind passing in a NULL for char *'s when you
give it an undef.  Although this is a feature in many cases, unfortunately
Heimdal will core if to give a NULL to krb5_parse_name() and so we must
check to ensure that we're not doing this.