Fix up tests to pass on last few commits.
We also specify a location for the sqlite DB in preference to the
system location so that the tests are a bit easier to run for those
that are not root.  We have not [yet] done this for the Kerberos DB,
though, so the tests are still not likely to pass for everyone.