Start using ECDH (Curve 25519) to negotiate keys.
We are replacing the old mechanisms used by krb5_keytab to create
and change keys with a method that uses ECDH to agree on a passwd
between both sides which is then used to generate keys of the
appropriate enctypes to install in the keytab and Kerberos DB.  In
order to ensure that these new methods are exclusively used by most
installations, we disable the older method by default and provide
a configuration parameter ($allow_fetch) to allow legacy behaviour.

We add unit tests for the new behaviour as well as adding a man
page (krb5_admind.conf.5) to describe the new configuration parameter.