First pass at the server side of initial host key bootstrapping support.
This is the first pass and is not expected to be complete.  We define a
number of new server side methods (which we do not yet export, we're just
testing).  create_bootstrap_id() which will create (using ECDH) a principal
which is named bootstrap with a randomly generated second component.
bind_host() which will create a binding between a host and a bootstrap id.
and bootstrap_host_key() which will allow a client to create a host key
if they are using the credentials that are bound to the host that they
are requesting.  The schema needed to be extended a little bit and we
also added basic tests into the test framework.