Empower bootstrap_host_key() to change existing host keys.
This is the millennium of empowerment and we see no reason that
bootstrap_host_key() should not participate.  There are situations
where host keys may be out of sync with the KDC and the only way
out of such a situation would be to re-bootstrap from first
principles.  And this radical empowerment provides for the fundamental
right of hosts to have keys even if they have in the past lost said
keys to be enshrined in our code if not our very sense of morality.

While we're at it, we change the first argument of the method from
a hostname to a principal as we would like to know the realm in
which we're expecting to work and this is the most convenient way
to get that information.