Use older kvnos when kinit'ing from a keytab if the newest one fails.
We change kinit_kt() to use older kvnos if newer ones fail.  The
basic algorithm is that we iterate over the supplied kt and put
entries into individual MEMORY: kts, one for each kvno.  We then
iterate backwards over the MEMORY: kts attempting to obtain initial
credentials.  We do this so that we can successfully kinit from a
kt which has new entries which were not committed to the KDC (which
can happen when we are rotating the host keys.)  At some point, we
should update Heimdal to do this natively as other applications
may have this issue during kt rotation.

We also restructured kinit_{anonymous,common,kt}().  kinit_common()
no longer makes much sense as kinit_kt() is too complex, so we
simply break kinit_{anonymous,kt}() into separate functions and
eliminate the code sharing.