Throw a more useful error in kinit_kt().
As we're looping trying again we need to handle errors slightly
differently than in the standard case with the K5BAIL() macro.  We
return the error we received from the first request (the highest
numbered kvno) under the assumption that it may be more useful.
Perhaps it would make more sense to catenate all of the errors
together?