Change Krb5Admin::genkeys() API.
We add an initial field to the genkeys() API which denotes the keys
changing operation which will later be invoked.  This is passed
via generate_ecdh_key1() to Krb5Admin::KerberosDB (either directly
or over the wire) along with with the principal in question and
generate_ecdh_key1() uses this information to validate that the
requested operation satisfies the requisite APIs before returning
the public key.  We do this so that operations which are disallowed
will fail before the newly generated keys are written into the
keytab rather than afterwards.  In order to do this, we had to
refactor a little code.