Use Kharon's ability to enforce ACLs.
This requires a little bit of a reorganisation of how the ACLs are
invoked.  We define the ACLs in scripts/krb5_admind, now.  We use
a Kharon::Entitlement::Or object to allow either the aclfile or
the self-service ACLs to be applied.  The self-service ACLs are
implemented using a Kharon::Entitlement::Object with the
Krb5Admin::KerberosDB object itself.  This means that the ACLs will
look for functions KHARON_COMMON_ACL and KHARON_ACL_<verb> in this
file and apply the permissions that they define.