Allow admin to extend the Krb5Admin::KerberosDB class in the config file.
We specify a new configuration variable $kmdb_class which if it is
defined will be used in preference to Krb5Admin::KerberosDB.  This
new class can be installed in the system perl5 locations, or it can
be simply defined in place in the configuration file as follows:

	$kmdb_class = 'MyKrb5Admin::KerberosDB';

	package MyKrb5Admin::KerberosDB;
	use base qw(Krb5Admin::KerberosDB);

	sub KHARON_ACL_fetch {

		return 'No.  No no no no no.  NO!';
	}

	1;

In general, this mechanism should only be used to override the ACLs
in Krb5Admin::KerberosDB as the changes are not visible on the
client side and hence new methods will not be visible to the client.
We shall define a mechanism more suitable for functionality extension
presently.