Fix impedence mismatch between generate_ecdh_key1 and create_bootstrap_id.
The arguments that generate_ecdh_key1's ACL function expects are
slightly different that create_bootstrap_id.  Since the the former's
ACLs call the latter's ACLs, we needed to solve this by sending a
list and failing to consume it.  While we are at it, we make the
ACL a little looser by only requiring that it is an anonymous
principal but not requiring that it belong to the realm that it
declares...