Preliminary ACL checking in test matrix.
This is not complete yet as we are mostly not checking the self-service
ACLs but rather just sacls but it is a start.  We do, however, check the
bootbindings near the end.  Interestingly, we run into problems if we
have more than one KerberosDB object in play near the end but we have
not analysed exactly what is going on there.