Fix issue with enctype == 0 and a bit of memory management.
Nico pointed out that the code did not behave properly in the case that
invalid enctypes were found.  Namely, if an enctype of zero was found,
things would go a bit wrong.  We took his suggested patch and extended
it to account for the Heimdal section of the #ifdef.

We also noted that the function leaked memory quite quickly and made a
couple of changes that turn this into a slow leak rather than a fast one.