In Krb5Admin::Krb5Host::Local::install_keys, don't fail if can't reach KDC.
We go back to the original logic from krb5_keytabc here as we need
to continue in case we are installing a bootstrap key or bootstrapping
a host key.  In other words, just because we failed to connect to
the KDC, it does not mean that we have failed at this point.