Implement Krb5Admin::C::need_new_key().
This function is intended to replace the functionality in krb5_keytabc
by the same name except it will use a more robust methodology.
Rather than attempting to kinit to the principal in question, it
rather performs a TGS_REQ for a ticket to service and then validates
it.  This is a closer proof to what we desire than that which we
had before.