Update t/krb5.conf to allow a KDC to be run on 127.0.0.1:1988.
Out with the leg warmers, it's 1988.  We update our test krb5.conf to
allow a KDC against the local database in preparation for testing
Krb5Admin::Krb5Host.  As we add 127.0.0.1:1988 to list of kdcs that
TEST.REALM supports, we must update getkdc.t to expect it.  We also
moving the log_file directive from [realms] TEST.REALM to the [kdc]
section where it belongs as a drive-by.