In Krb5Admin::Krb5Host::Local, define an option ``testing''.
If set to true, this will remove a few chmod(2)s, chown(2)s and
stat(2)s to allow the test suite to run as a local user.  The code
naturally assumes that it is being run as root as it must manipulate
the keytabs of mutiple users and must necessarily change the
ownership of them.  This is inconvenient for testing, however, so
we allow this behaviour to be conditionally disabled.