Improve key installation logic in Krb5Admin::Krb5Host::Local.
We improve our ability to detect if we need to install new keys by
performing a TGS for the service principal and comparing it to the
contents of the keytab rather than performing an AS_REQ with the
keytab.  This is good for a number of reasons not least of which is
that we may install service principals in such a way that AS_REQs
are disallowed.

We also update our logic in install_key() when we find that we have
keys.  We no longer simply compare the kvno that the KDC reports
to what we have in the keytab but rather perform the above test to
ensure that the key we have is actually correct and also has the
correct kvno.  Given that we might get a service ticket from a
slave, we compare the kvno to what we expect and rerun the test
for a number of seconds if the kvno is too low.  (There are still
a few issues with this logic which have mostly been pointed out in
XXXrcd comments.)

We update the test script to test that faulty keys will cause the
keys to be renogotiated in a simple case.