Krb5Admin::Krb5Host::Local: remove undef enctypes when writing keys.
When writing a new set of keys into a keytab, we need to remove
extraneous enctypes.  What we do is if write_kt() is writing at
least one key matching (princ, kvno), we assume that it is writing
all enctypes for (princ, kvno) and remove all of the enctypes that
are not passed in.  This is important as we can run into situations
where the next kvno for a princ has already had keys written but
for a different library setting and that transaction was not
completed against the KDC.  The next write will overwrite the same
kvno but (until this change) it would not clobber the entirety of
the previous write.