Add HKDF support to Krb5Admin::C.
We add HKDF support using the code provided in IETF RFC 6234.  Our
approach was to format the RFC slightly, add few lines denoting the
beginning and end of each function, and write a Perl script which
extracts the various source files from the RFC.

We also add a bit of logic to the Swig bits to generate some bindings
to make it easy to use from Perl.