CURVE_25519: fix a few things.
We redefine curve25519_start to return a list ref rather than a list
to make exporting it via Kharon a little easier.

We rename $privdata -> $priv globally.  This just looks a little nicer.

We change do_nway() to talk to the last host first which we presume
is the KDC.  We also write a large comment stating that we need to
carefully examine the ordering of these various conversations when
we begin to evaluate locking.  This should be relatively soon.