First pass at using CURVE25519_NWAY to negotiate keys.
We're limiting ourselves at this first pass to simply replacing the
existing negotiation with the way defined in our new module.  This
is not compatible with the current mechanism.  The new server will
support the old mechanism, though, so the upgrade strategy will be
to simply upgrade the KDC before the hosts.