Change CURVE25519_NWAY's order of processing.
In order to properly deal with writing and locking, we need to have
CURVE25519_NWAY process its host list in a deterministic and specific
order.  We will call curve25519_start() on the host list in order and
call curve25519_final() in reverse order.  This allows the locking,
unlocking, and writing to proceed appropriately.