Update API for CURVE25519_NWAY to allow in flight $priv mods.
We update the API for CURVE25519_NWAY::do_nway() to take %opts and
define an option to pass in called ``privfunc''.  This function can
be by the participants to modify $priv in flight subject to certain
constraints.  curve25519_start() now returns a list ref containing
an extra element.  The last element is arbitrary and is defined by
the module which inherits this class.  If a privfunc is passed in
then this arbitrary element is passed to privfunc and if privfunc
returns a defined value then that value will replace $priv.

We also return $priv from do_nway() so that the caller can examine
the resultant $priv to construct return values, et al.