Use the new CURVE25519_NWAY $priv modification to handle kvnos better.
We note that currently we are going to start with kvno == 2 again
but we'd rather start with 1 at some point.  What we were doing
before, though, required the client to know what the kvno was going
to be before it communicated with the KDC.  This isn't optimal as
it would stop us from optimising the procedure in the future.  So,
we have switched to a procedure where the KDC will set the kvno as
part of the ECDH negotiation and the clients will honour it.