Krb5Admin::Krb5Host::Local.pm: No need to require -c for clusters.
Our older code would output an error if running an unadorned krb5_keytab
for a cluster principal if the keys needed to be updated.  Given that
we now synchronise the updates within a cluster, this should no longer
be necessary and so we remove the error allowing the code to continue.