Krb5Admin::Krb5Host::Local: put prestashed tickets in realm-based paths
We note that if you obtain tickets for different principals with
the same localname but different realms, well, they need to be in
different places.  We make the simplifying assumption that the
realm will be the only difference and then say that user@REALM will
prestash into:

	/var/spool/tickets/REALM/user

if the realm isn't the default realm.  We also create a symlink:

	/var/spool/tickets/DEFAULT_REALM -> .

so that all prestashed tickets can be referred to using the realm.