Krb5Admin::Daemon: we start to reuse the Krb5Admin::KerberosDB
in preforked servers.  This involved a few changes to Krb5Admin::KerberosDB
as well.  Basically, we make sure that when we ->set_creds on our
Kharon::Entitlements objects that the change re-opens the HDB and
the like.  And then we restructured Krb5Admin::Daemon to do that
rather than destroying and recreating the object.

This solves an fd leak which occurs in the old code because we had
created a circular dependency and the Krb5Admin::KerberosDB's were
not being cleaned up and so were keeping their HDB and SQLite3 fds
open.